cleaning and refactoring

+++ app/views.py
@@ -22,20 +22,17 @@ from app.utils import database, interactions
 
 @login_required
 def home(request):
-    """Home page"""
-
-    if "delete" in request.POST:
-        Media.objects.get(
-            media_id=request.POST["delete"],
-            user=request.user,
-            api=request.POST["api"],
-        ).delete()
-        return redirect("home")
-
-    elif "status" in request.POST:
-        if request.user.is_authenticated:
+    if request.method == "POST":
+        if "delete" in request.POST:
+            Media.objects.get(
+                media_id=request.POST["delete"],
+                user=request.user,
+                api=request.POST["api"],
+            ).delete()
+        elif "status" in request.POST:
             database.edit_media(request)
-            return redirect("home")
+
+        return redirect("home")
 
     queryset = Media.objects.filter(user_id=request.user)
     data = {
@@ -44,7 +41,6 @@ def home(request):
         "anime": {"media": [], "statuses": {}},
         "manga": {"media": [], "statuses": {}},
     }
-
     for media in queryset:
         media_type = media.media_type
         data[media_type]["media"].append(media)
@@ -52,68 +48,53 @@ def home(request):
         if status not in data[media_type]["statuses"]:
             data[media_type]["statuses"][status] = []
         data[media_type]["statuses"][status].append(media)
+    return render(request, "app/home.html", {"data": data})
 
-    return render(
-        request,
-        "app/home.html",
-        {
-            "data": data,
-        },
-    )
 
 @login_required
 def search(request):
-    """Search page"""
-
     api = request.GET.get("api")
     query = request.GET.get("q")
+    request.session['last_selected_api'] = api
 
-    if "delete" in request.POST:
-        Media.objects.get(
-            media_id=request.POST["delete"],
-            user=request.user,
-            api=request.POST["api"],
-        ).delete()
-        return redirect("/search?api=" + api + "&q=" + query)
-
-    elif "status" in request.POST:
-        if request.user.is_authenticated:
-            if Media.objects.filter(
-                media_id=request.POST["media_id"],
+    if request.method == "POST":
+        if "delete" in request.POST:
+            Media.objects.get(
+                media_id=request.POST["delete"],
                 user=request.user,
                 api=request.POST["api"],
-            ).exists():
-                database.edit_media(request)
+            ).delete()
+        elif "status" in request.POST:
+            if request.user.is_authenticated:
+                if Media.objects.filter(
+                    media_id=request.POST["media_id"],
+                    user=request.user,
+                    api=request.POST["api"],
+                ).exists():
+                    database.edit_media(request)
+                else:
+                    database.add_media(request)
             else:
-                database.add_media(request)
-        else:
-            messages.error(request, "Log in is required to track media.")
-            return redirect("login")
-
+                messages.error(request, "Log in is required to track media.")
+                return redirect("login")
         return redirect("/search?api=" + api + "&q=" + query)
 
     query_list = interactions.search(api, query)
     context = {"query_list": query_list}
-
     return render(request, "app/search.html", context)
 
 
 def register(request):
-
-    if "username" in request.POST:
-        form = UserRegisterForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect("login")
-    else:
-        form = UserRegisterForm()
+    form = UserRegisterForm(request.POST if request.method == "POST" else None)
+    if form.is_valid():
+        form.save()
+        return redirect("login")
     return render(request, "app/register.html", {"form": form})
 
 
 def login(request):
-
-    if "username" in request.POST:
-        form = AuthenticationForm()
+    form = AuthenticationForm()
+    if request.method == "POST":
         user = authenticate(
             request,
             username=request.POST["username"],
@@ -127,81 +108,67 @@ def login(request):
                 request,
                 "Please enter a correct username and password. Note that both fields may be case-sensitive.",
             )
-    else:
-        form = AuthenticationForm()
-
     return render(request, "app/login.html", {"form": form})
 
 
 @login_required
 def profile(request):
-    if "username" in request.POST:
-        user_form = UserUpdateForm(request.POST, instance=request.user)
+    user_form = UserUpdateForm(request.POST or None, instance=request.user)
+    password_form = PasswordChangeForm(request.user, request.POST or None)
+    
+    if request.method == 'POST':
         if user_form.is_valid():
             user_form.save()
             messages.success(request, "Your account has been updated!")
-            return redirect("profile")
 
-    elif "new_password1" in request.POST:
-        password_form = PasswordChangeForm(request.user, request.POST)
-        if password_form.is_valid():
+        elif password_form.is_valid():
             password = password_form.save()
             update_session_auth_hash(request, password)
             messages.success(request, "Your password has been updated!")
-            return redirect("profile")
 
-    elif request.POST.get("mal") and request.POST.get("mal-btn"):
-        if interactions.import_myanimelist(request.POST.get("mal"), request.user):
-            messages.success(request, "Your MyAnimeList has been imported!")
-        else:
-            messages.error(request, "User not found")
+        elif request.POST.get("mal"):
+            if interactions.import_myanimelist(request.POST.get("mal"), request.user):
+                messages.success(request, "Your MyAnimeList has been imported!")
+            else:
+                messages.error(request, "User not found")
+
+        elif request.FILES.get("tmdb"):
+            if interactions.import_tmdb(request.FILES.get("tmdb"), request.user):
+                messages.success(request, "Your TMDB list has been imported!")
+            else:
+                messages.error(
+                    request,
+                    'Error importing your list, make sure it\'s a CSV file containing the word "ratings" or "watchlist" in the name'
+                )
+
+        elif request.POST.get("anilist"):
+            errors = interactions.import_anilist(request.POST.get("anilist"), request.user)
+            if len(errors) == 0:
+                messages.success(request, "Your AniList has been imported!")
+            elif errors[0] == "User not found":
+                messages.error(request, "User not found")
+            else:
+                message = "<br/>".join(errors)
+                messages.error(
+                    request,
+                    format_html(
+                        "<b>{}</b> <br>{}",
+                        "Couldn't find a matching MAL ID for: ",
+                        mark_safe(message),
+                    ),
+                )
         return redirect("profile")
-
-    elif request.FILES.get("tmdb") and request.POST.get("tmdb-btn"):
-        if interactions.import_tmdb(request.FILES.get("tmdb"), request.user):
-            messages.success(request, "Your TMDB list has been imported!")
-        else:
-            messages.error(
-                request,
-                'Error importing your list, make sure it\'s a CSV file containing the word "ratings" or "watchlist" in the name'
-            )
-
-        return redirect("profile")
-
-    elif request.POST.get("anilist") and request.POST.get("anilist-btn"):
-
-        errors = interactions.import_anilist(request.POST.get("anilist"), request.user)
-        if len(errors) == 0:
-            messages.success(request, "Your AniList has been imported!")
-        elif errors[0] == "User not found":
-            messages.error(request, "User not found")
-        else:
-            message = "<br/>".join(errors)
-            messages.error(
-                request,
-                format_html(
-                    "<b>{}</b> <br>{}",
-                    "Couldn't find a matching MAL ID for: ",
-                    mark_safe(message),
-                ),
-            )
-
-        return redirect("profile")
-
-    else:
-        user_form = UserUpdateForm(instance=request.user)
-        password_form = PasswordChangeForm(request.user)
 
     context = {"user_form": user_form, "password_form": password_form}
-
     return render(request, "app/profile.html", context)
+
 
 
 def edit(request, media_type, media_id):
 
-    if media_type == "anime" or media_type == "manga":
+    if media_type in ["anime", "manga"]:
         media = interactions.mal_edit(request, media_type, media_id)
-    elif media_type == "movie" or media_type == "tv":
+    elif media_type in ["movie", "tv"]:
         media = interactions.tmdb_edit(request, media_type, media_id)
 
     data = {'title': media["response"]["title"], 'year': media["response"]["year"], 'media_type': media["response"]["media_type"]}
